Refactor: Modularize probes and enhance HTTP strategy
- Reorganize host probes into separate packages (cpu, memory, network, logicaldisk)
- Implement ConfigSource interface for unified configuration access
- Add auto-update functionality in offline mode with configurable settings
- Enhance HTTP strategy with improved tag filtering and URL handling
- Fix test failures after refactoring to use configSource
- Update user guide integration with enhanced markdown rendering
- Add cache configuration management with dynamic TTL updates
- Improve language consistency across documentation and comments

+++ internal/agent/validators/validators_test.go
@@ -42,8 +42,8 @@ func TestIsDuration(t *testing.T) {
 		want  bool
 	}{
 		{
-			name:  "Float64",
-			value: float64(10),
+			name:  "Int",
+			value: 10,
 			want:  true,
 		},
 		{
@@ -58,7 +58,7 @@ func TestIsDuration(t *testing.T) {
 		},
 		{
 			name:  "Invalid Type",
-			value: 10,
+			value: float64(10),
 			want:  false,
 		},
 	}
